:sparkles: Implemented ViewModel and fetch a list of beers from Api

+++ app/src/main/java/com/example/beerrecipes/domain/BeerDto.kt
@@ -2,7 +2,7 @@ package com.example.beerrecipes.domain
 
 data class BeerDto(
     val abv: Double,
-    val attenuation_level: Int,
+    val attenuation_level: Double,
     val boil_volume: BoilVolume,
     val brewers_tips: String,
     val contributed_by: String,
@@ -10,17 +10,17 @@ data class BeerDto(
     val ebc: Int,
     val first_brewed: String,
     val food_pairing: List<String>,
-    val ibu: Int,
+    val ibu: Double,
     val id: Int,
     val image_url: String,
     val ingredients: Ingredients,
     val method: Method,
     val name: String,
     val ph: Double,
-    val srm: Int,
+    val srm: Double,
     val tagline: String,
     val target_fg: Int,
-    val target_og: Int,
+    val target_og: Double,
     val volume: Volume
 )
